refactor(feedback-widget): final cleanup — drop unused imports and split test file
Orchestrator cleanup:
- Drop unused imports: getInitials/avatarColor/normalizeReviewStatus/
  timeAgo from format, popoverRef (no longer referenced after the
  CommentInputPopover extraction took ownership of its own ref).

Test split:
- The 1200-line src/__tests__/FeedbackWidget.test.tsx is broken into
  7 focused files under src/__tests__/feedbackWidget/, sharing the
  mockFetch/seedComment/commentsResponse/widgetTestSetup/
  enterCommentingMode helpers via src/__tests__/helpers/feedbackWidgetHarness.ts:

  - lifecycle.test.tsx — mount/fetch/encoding/PATCH/instruction-bar
  - submit.test.tsx — POST flow, screenshot preview, dedup, scrim cancel,
    avatar-name editor, failed POST guard
  - sidebar.test.tsx — sidebar X close, inline edit, filter empty state,
    detail-popover trigger, pin numbering
  - pin.test.tsx — pin marker hover/click/edit/cancel/save/delete and
    scrim clear through the orchestrator's CommentPin callbacks
  - pill.test.tsx — floating pill button taps + Shift+A agent reveal
  - shortcuts.test.tsx — c/s/m/h shortcuts + every Escape cascade level
  - e2e.test.tsx — full submit-and-approve walk-through

Each file under 230 lines. Each owns its `vi.mock('../../lib/screenshotCapture')`
inline (vi.mock factories are hoisted per-file, can't be shared via import).
